# Skip no-op adjustments in trans